+++ src/stats/hexbin.jl
@@ -7,6 +7,7 @@ Plots a heatmap with hexagonal bins for the observations `xs` and `ys`.
 
 ### Specific to `Hexbin`
 
+- `weights = nothing`: Weights for each observation.  Can be `nothing` (each observation carries weight 1) or any `AbstractVector{<: Real}` or `StatsBase.AbstractWeights`. 
 - `bins = 20`: If an `Int`, sets the number of bins in x and y direction. If a `Tuple{Int, Int}`, sets the number of bins for x and y separately.
 - `cellsize = nothing`: If a `Real`, makes equally-sided hexagons with width `cellsize`. If a `Tuple{Real, Real}` specifies hexagon width and height separately.
 - `threshold::Int = 1`: The minimal number of observations in the bin to be shown. If 0, all zero-count hexagons fitting into the data limits will be shown.
@@ -21,12 +22,14 @@ Plots a heatmap with hexagonal bins for the observations `xs` and `ys`.
     return Attributes(;
                       colormap=theme(scene, :colormap),
                       colorrange=Makie.automatic,
+                      weights=nothing,
                       bins=20,
                       cellsize=nothing,
                       threshold=1,
                       scale=identity,
                       strokewidth=0,
-                      strokecolor=:black)
+                      strokecolor=:black,
+        )
 end
 
 function spacings_offsets_nbins(bins::Tuple{Int,Int}, cellsize::Nothing, xmi, xma, ymi, yma)
@@ -69,6 +72,10 @@ function data_limits(hb::Hexbin)
     return Rect3f(no, nw)
 end
 
+get_weight(weights, i) = Float64(weights[i])
+get_weight(::StatsBase.UnitWeights, i) = 1e0
+get_weight(::Nothing, i) = 1e0
+
 function Makie.plot!(hb::Hexbin{<:Tuple{<:AbstractVector{<:Point2}}})
     xy = hb[1]
 
@@ -76,7 +83,7 @@ function Makie.plot!(hb::Hexbin{<:Tuple{<:AbstractVector{<:Point2}}})
     count_hex = Observable(Float64[])
     markersize = Observable(Vec2f(1, 1))
 
-    function calculate_grid(xy, bins, cellsize, threshold, scale)
+    function calculate_grid(xy, weights, bins, cellsize, threshold, scale)
         empty!(points[])
         empty!(count_hex[])
 
@@ -102,13 +109,14 @@ function Makie.plot!(hb::Hexbin{<:Tuple{<:AbstractVector{<:Point2}}})
         xsize = xspacing * 2
         rx = xsize / sqrt3
 
-        d = Dict{Tuple{Int,Int},Int}()
+        d = Dict{Tuple{Int,Int}, Float64}()
 
         # for the distance measurement, the y dimension must be weighted relative to the x
         # dimension according to the different sizes in each, otherwise the attribution to hexagonal
         # cells is wrong
         yweight = xsize / ysize
 
+        i = 1
         for (_x, _y) in xy
             nx, nxs, dvx = nearest_center(_x, xspacing, xoff)
             ny, nys, dvy = nearest_center(_y, yspacing, yoff)
@@ -132,7 +140,8 @@ function Makie.plot!(hb::Hexbin{<:Tuple{<:AbstractVector{<:Point2}}})
                 )
             end
 
-            d[id] = get(d, id, 0) + 1
+            d[id] = get(d, id, 0) + (get_weight(weights, i))
+            i += 1
         end
 
         if threshold == 0
@@ -141,7 +150,7 @@ function Makie.plot!(hb::Hexbin{<:Tuple{<:AbstractVector{<:Point2}}})
                 for ix in 0:_nx-1
                     _x = xoff + 2 * ix * xspacing + (isodd(iy) * xspacing)
                     _y = yoff + iy * yspacing
-                    c = get(d, (ix, iy), 0)
+                    c = get(d, (ix, iy), 0e0)
                     push!(points[], Point2f(_x, _y))
                     push!(count_hex[], scale(c))
                 end
@@ -162,7 +171,7 @@ function Makie.plot!(hb::Hexbin{<:Tuple{<:AbstractVector{<:Point2}}})
         notify(points)
         return notify(count_hex)
     end
-    onany(calculate_grid, xy, hb.bins, hb.cellsize, hb.threshold, hb.scale)
+    onany(calculate_grid, xy, hb.weights, hb.bins, hb.cellsize, hb.threshold, hb.scale)
     # trigger once
     notify(hb.bins)